JVM: do not write trivial SMAPs to classes outside inline funs
where trivial == those that map the file to itself.

Also, fix a bug where running a kts file that throws an exception will
display "Unknown source" for all lines corresponding to the top level
class if there are no inline calls in it; e.g.

    fun f() {
        throw RuntimeException("...")
    }
    f()

used to output

    java.lang.RuntimeException: ...
        at X.f(Unknown Source)
        at X.<init>(Unknown Source)

instead of

    java.lang.RuntimeException: ...
        at X.f(x.kts:2)
        at X.<init>(x.kts:4)

Similarly, you can now see which line of a REPL expression has failed:

    >>> fun f() {
    ...   throw RuntimeException("...")
    ... }
    >>> f()
    java.lang.RuntimeException: ...
        at Line_2.f(Line_2.kts:2)

(Note that Line_2 here refers to some internal index, not to the actual
line 2 of f; writing more expressions will cause this index to increase.)